# back-end/Documentation/TechDon REST API - OpenAPI.yaml
...
openapi: 3.0.1
info:
  title: TechDon
  description: This is the REST API documentation for the end points for TechDon
    platform will allow students to sign up, register for items and receive them in
    a simple, easy and straight forward procedure. The main aim of the platform is
    to provide students from a less privileged backgrounds to be able to apply for
    electronics to be donated which in turn will assist them in their educational
    studies. We will integrate a first come first served online que system where students
    can apply and wait there turn to be donated electrical goods. Our purpose behind
    this procedure is to insure a fair and easy process o deciding who justifies to
    be given the device first.  We will incorporate an ID check system where students
    must upload their id to ensure they are eligible to receive the device. Donators
    can be an individual or a company looking to help better the quality of education
    of students. This will be done by them uploading pictures of the device which
    later found student to be donated to , contacts will be exchanged to be able to
    finalize the donation , if it’s a in person or shipped delivery.
  termsOfService: http://localhost:8080
  contact:
    email: 2026156@Brunel.ac.uk
  license:
    name: Thoybur Rohman / Philipp Bibik
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.1.0
servers:
- url: https://Group50/v1
...
paths:
...
  /donor/device/new:
    post:
      tags:
      - donor
      summary: List a new device for donation
      description: Endpoint for listing a new device for donation on the platform. Requires an authToken of a donor account and the required fields to by supplied. Device description and photos are optional.
      operationId: UploadItem
      parameters:
      - name: authToken
        in: query
        description: Auth token to prove identity
        required: true
        schema:
          type: string
      - name: productType
        in: query
        description: Type of donated product (laptop / monitor / etc.)
        required: true
        schema:
          type: string
      - name: productName
        in: query
        description: Name of the product
        required: true
        schema:
          type: string
      - name: productLocation
        in: query
        description: Location of product (City or Postcode)
        required: true
        schema:
          type: string
      - name: productDescription
        in: query
        description: Optional description of the product
        schema:
          type: string
      responses:
        200:
          description: Device uploaded successfully
        400:
          description: Invalid parameters
        401:
          description: Could not authenticate
  /donor/device/listedDevices:
    get:
      tags:
      - donor
      summary: Get list of donated devices
      description: This endpoint returns a list of all devices listed by a donor as a
        String Array ["0b52b9a5-0737-4762-b628-b55eb988744c", "6920289b-d447-4fdb-b178-03e4562a9839", "3127fec8-5bc3-4ac8-b831-697e831581dc", ...]
      operationId: getDevices
      parameters:
      - name: authToken
        in: query
        description: Auth token for authentication
        required: true
        schema:
          type: string
      responses:
        200:
          description: List of device IDs which belong to user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/device'
        400:
          description: Invalid parameters
        401:
          description: Could not authenticate
  /donor/device/load:
    get:
      tags:
      - donor
      summary: Load information about a donated device
      description: Endpoint for donors to load information about a device which has been donated by them. Expects an auth token and device id
      operationId: loadDevices
      parameters:
      - name: authToken
        in: query
        description: Auth token for authentication
        required: true
        schema:
          type: string
      - name: deviceId
        in: query
        description: ID of device to be loaded
        required: true
        schema:
          type: string
      responses:
        200:
          description: Device data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getDevice'
        400:
          description: Invalid parameters
        401:
          description: Could not authenticate
  /donor/device/remove:
    delete:
      tags:
      - donor
      summary: Remove donated device
      description: This endpoint is for donor to remove devices they have listed from the platform. 
      operationId: DeviceRemove
      parameters:
      - name: authToken
        in: query
        description: Auth token for authentication
        required: true
        schema:
          type: string
      - name: deviceId
        in: query
        description: ID of device to be removed
        required: true
        schema:
          type: string
      responses:
        200:
          description: Device succfully removed
        400:
          description: Invalid parameters
        401:
          description: Could not authenticate
  /donor/device/update:
    post:
      tags:
      - donor
      summary: Update device info
      description: Endpoint for updating the information of a device. Requires the authToken and deviceId to specify which device is to be updated and prove authorization, along with any number of device fields and the new values they should be set to. Unspecified values are not to be changed. Returns the new device state.
      operationId: Update
      parameters:
      - name: authToken
        in: query
        description: Auth token for authentication
        required: true
        schema:
          type: string
      - name: deviceId
        in: query
        description: ID of device to be removed
        required: true
        schema:
          type: string
      - name: productType
        in: query
        description: Type of donated product (laptop / monitor / etc.)
        schema:
          type: string
      - name: productName
        in: query
        description: Name of the product
        schema:
          type: string
      - name: productLocation
        in: query
        description: Location of product (City or Postcode)
        schema:
          type: string
      - name: productDescription
        in: query
        description: Optional description of the product
        schema:
          type: string
      responses:
        200:
          description: Device updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getDevice'
        400:
          description: Invalid parameters
        401:
          description: Could not authenticate
...
components:
  schemas:
...
    getDevice:
      type: object
      properties:
        deviceType:
          type: string
          format: string
          example: laptop
        deviceName:
          type: string
          example: HP 14in Ryzen 3 4GB 128GB FHD Laptop - Black
        deviceLocation:
          type: string
          example: London
        deviceDescription:
          type: string
          example: Slightly used, but in good condition. Collection only.
...
    device:
      type: array
      description: 'List of device IDs'
      example:
      - "0b52b9a5-0737-4762-b628-b55eb988744c"
      - "6920289b-d447-4fdb-b178-03e4562a9839"
      - "3127fec8-5bc3-4ac8-b831-697e831581dc"
      items:
        type: string